pyproject.toml
[project]
authors = ['José Roberto Ferreira']
description = "X3"
name = "X3"
requires-python = ">=3.12"
version = "0.1.0"
dependencies = [
  "annotated-types==0.7.0",
  "anyio==4.9.0",
  "click==8.1.8",
  "colorama==0.4.6",
  "fastapi==0.115.12",
  "graphql-core==3.2.6",
  "greenlet==3.2.0",
  "h11==0.14.0",
  "httptools==0.6.4",
  "idna==3.10",
  "packaging==25.0",
  "pydantic==2.11.3",
  "pydantic_core==2.33.1",
  "pyodbc==5.2.0",
  "python-dateutil==2.9.0.post0",
  "python-decouple==3.8",
  "python-dotenv==1.1.0",
  "PyYAML==6.0.2",
  "six==1.17.0",
  "sniffio==1.3.1",
  "SQLAlchemy==2.0.40",
  "sqlmodel==0.0.24",
  "starlette==0.46.2",
  "strawberry-graphql==0.266.0",
  "typing-inspection==0.4.0",
  "typing_extensions==4.13.2",
  "uvicorn==0.34.2",
  "watchfiles==1.0.5",
  "websockets==15.0.1"
]

[project.optional-dependencies]
dev = [
    "inflect==7.5.0",
    "more-itertools==10.7.0",
    "mslex==1.3.0",
    "psutil==6.1.1",
    "sqlacodegen==3.0.0",
    "taskipy==1.14.1",
    "tomli==2.2.1",
    "typeguard==4.4.2"
]

[tool.taskipy.tasks]
run = "uvicorn app.main:app --reload"

[tool.ruff]
line-length = 120
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'
